requirements.txt
pyexiftool>=0.5.6
pillow>=11.2.1
python-dateutil>=2.9.0
pytest>=8.4.0
